pyproject.toml
[project]
name = "pandocster"
version = "2.2.0"
readme = "README.md"
license = {file = "LICENSE.txt"}
authors = [
    {name = "Leo Sadovsky", email = "leo@sadovskie.ru"}
]
keywords = ["pandoc", "pandocster", "pandoc-filters", "doc-as-code", "documentation", "markdown", "pdf", "html", "docx"]
description = "Pandocster: a CLI helper around pandoc workflows, it helps you to build your documentation from markdown files."
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5.0",
    "mypy>=1.9",
    "pre-commit>=3.0",
    "types-click>=7.1.8",
]

[project.scripts]
pandocster = "cli.entrypoint:main"

[build-system]
requires = ["setuptools>=61", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"pandoc-filters" = ["*.lua"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-ra --strict-markers"

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.10"
mypy_path = "src"
packages = ["pandocster"]
strict_optional = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

